Turn off previous note when stop playing
I've noticed occasional stuck notes using this example. This change seems to fix the problem. The problem, I suspect, happens when the *on* thumb is released at the same time as there are other fingering changes, in which case this sketch would turn off the note corresponding to the new fingering instead of turning off the note that was previously turned on, resulting in a stuck note.